+++ client/src/components/popup/erase.tsx
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux'
 import { useAppDispatch } from '../../hooks'
 import { deleteProgress, selectProgress } from '../../state/progress'
 import { downloadFile } from '../world_tree'
-import { Button } from '../button'
+import { Button } from '../utils'
 import { Trans, useTranslation } from 'react-i18next'
 import { useContext } from 'react'
 import { PopupContext } from './popup'
