+++ client/src/components/popup/upload.tsx
@@ -6,7 +6,7 @@ import { useSelector } from 'react-redux'
 import { useAppDispatch } from '../../hooks'
 import { GameProgressState, loadProgress, selectProgress } from '../../state/progress'
 import { downloadFile } from '../world_tree'
-import { Button } from '../button'
+import { Button } from '../utils'
 import { Trans, useTranslation } from 'react-i18next'
 import { PopupContext } from './popup'
 import { useContext } from 'react'
